move cast_simplify to ir_simplify
